Added extra warnings and RTC stuff for MSVC

+++ cpp/setup.cpp
@@ -6,11 +6,12 @@ namespace {
 uint16_t port = 0;
 }
 
-int get_random_port() {
+uint16_t get_random_port() {
   if (port == 0) {
     std::random_device rd;
     std::uniform_int_distribution<int> dist(0, 16383);
-    port = 49152 + dist(rd); // Use ports in the range 49152-65535
+    port = 49152 + static_cast<uint16_t>(dist(rd)); // Use ports in the range
+                                                    // 49152-65535
   }
   return port;
 }
